Move colors to run-support project
