+++ framework/src/sbt-plugin/src/main/scala/play/sbtplugin/Colors.scala
@@ -5,25 +5,17 @@ package play.sbtplugin
 
 object Colors {
 
-  import scala.Console._
+  import play.runsupport.{ Colors => RunColors }
 
-  lazy val isANSISupported = {
-    Option(System.getProperty("sbt.log.noformat")).map(_ != "true").orElse {
-      Option(System.getProperty("os.name"))
-        .map(_.toLowerCase)
-        .filter(_.contains("windows"))
-        .map(_ => false)
-    }.getOrElse(true)
-  }
+  lazy val isANSISupported = RunColors.isANSISupported
 
-  def red(str: String): String = if (isANSISupported) (RED + str + RESET) else str
-  def blue(str: String): String = if (isANSISupported) (BLUE + str + RESET) else str
-  def cyan(str: String): String = if (isANSISupported) (CYAN + str + RESET) else str
-  def green(str: String): String = if (isANSISupported) (GREEN + str + RESET) else str
-  def magenta(str: String): String = if (isANSISupported) (MAGENTA + str + RESET) else str
-  def white(str: String): String = if (isANSISupported) (WHITE + str + RESET) else str
-  def black(str: String): String = if (isANSISupported) (BLACK + str + RESET) else str
-  def yellow(str: String): String = if (isANSISupported) (YELLOW + str + RESET) else str
+  def red(str: String): String = RunColors.red(str)
+  def blue(str: String): String = RunColors.blue(str)
+  def cyan(str: String): String = RunColors.cyan(str)
+  def green(str: String): String = RunColors.green(str)
+  def magenta(str: String): String = RunColors.magenta(str)
+  def white(str: String): String = RunColors.white(str)
+  def black(str: String): String = RunColors.black(str)
+  def yellow(str: String): String = RunColors.yellow(str)
 
 }
-
